test delete quand objet utiliser

+++ src/Controller/MaterialController.php
@@ -18,7 +18,7 @@ class MaterialController extends AbstractController
     #[Route('/', name: 'app_material_index', methods: ['GET'])]
     public function index(MaterialRepository $materialRepository): Response
     {
-       
+
         return $this->render('material/index.html.twig', [
             'materials' => $materialRepository->findAll(),
         ]);
@@ -28,8 +28,8 @@ class MaterialController extends AbstractController
     #[Route('/new', name: 'app_material_new', methods: ['GET', 'POST'])]
     public function new(Request $request, MaterialRepository $materialRepository): Response
     {
-       
-        
+
+
         $material = new Material();
         $form = $this->createForm(MaterialType::class, $material);
         $form->handleRequest($request); // recupere la requete si lors de lenvoi du forms
@@ -66,7 +66,7 @@ class MaterialController extends AbstractController
     #[Route('/{id}/edit', name: 'app_material_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Material $material, MaterialRepository $materialRepository): Response
     {
-        
+
         $form = $this->createForm(MaterialType::class, $material);
         $form->handleRequest($request);
 
@@ -84,12 +84,19 @@ class MaterialController extends AbstractController
 
 
 
-    
+
     #[Route('/{id}', name: 'app_material_delete', methods: ['POST'])]
     public function delete(Request $request, Material $material, MaterialRepository $materialRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$material->getId(), $request->request->get('_token'))) {
-            $materialRepository->remove($material, true);
+        $reservations = $material->getReservations();
+
+        if ($reservations->count() > 0) {
+            $this->addFlash('error', 'Ce materiel est actuellement utilisé');
+        } else {
+            if ($this->isCsrfTokenValid('delete' . $material->getId(), $request->request->get('_token'))) {
+                $materialRepository->remove($material, true);
+                $this->addFlash('success', 'Material deleted successfully.');
+            }
         }
 
         return $this->redirectToRoute('app_material_index', [], Response::HTTP_SEE_OTHER);
